Split waveform edit fade drag logic
